# docs/openapi-semana9.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MiniMarket Plus API",
    "description": "API REST de MiniMarket Plus para autenticación, catálogo, carritos, inventario, ventas, sucursales, stock por sucursal, proveedores, órdenes de compra, pedidos, promociones y reportes. Protegida mediante JWT y enriquecida con enlaces HATEOAS.",
    "contact": {
      "name": "Equipo MiniMarket Plus"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/sucursales/{id}": {
      "get": {
        "tags": [
          "Sucursales"
        ],
        "operationId": "obtener",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la sucursal",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Sucursal encontrada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelSucursal"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          },
          "404": {
            "description": "Sucursal no encontrada"
          }
        },
        "summary": "Obtener sucursal",
        "description": "Obtiene una sucursal por su identificador."
      },
      "put": {
        "tags": [
          "Sucursales"
        ],
        "operationId": "actualizar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la sucursal",
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sucursal"
              }
            }
          },
          "required": true,
          "description": "Datos actualizados de la sucursal"
        },
        "responses": {
          "200": {
            "description": "Sucursal actualizada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelSucursal"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Sucursal no encontrada"
          },
          "409": {
            "description": "Nombre de sucursal duplicado"
          }
        },
        "summary": "Actualizar sucursal",
        "description": "Actualiza los datos de una sucursal existente. Requiere ROLE_ADMIN."
      },
      "delete": {
        "tags": [
          "Sucursales"
        ],
        "operationId": "eliminar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la sucursal",
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Sucursal eliminada"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Sucursal no encontrada"
          },
          "409": {
            "description": "La sucursal posee relaciones activas"
          }
        },
        "summary": "Eliminar sucursal",
        "description": "Elimina una sucursal existente. Requiere ROLE_ADMIN."
      }
    },
    "/api/proveedores/{id}": {
      "get": {
        "tags": [
          "Proveedores"
        ],
        "operationId": "obtener_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del proveedor",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Proveedor encontrado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProveedor"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Proveedor no encontrado"
          }
        },
        "summary": "Obtener proveedor",
        "description": "Obtiene un proveedor por su identificador. Requiere ROLE_ADMIN."
      },
      "put": {
        "tags": [
          "Proveedores"
        ],
        "operationId": "actualizar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del proveedor",
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Proveedor"
              }
            }
          },
          "required": true,
          "description": "Datos actualizados del proveedor"
        },
        "responses": {
          "200": {
            "description": "Proveedor actualizado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProveedor"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Proveedor no encontrado"
          },
          "409": {
            "description": "Correo de proveedor duplicado"
          }
        },
        "summary": "Actualizar proveedor",
        "description": "Actualiza un proveedor existente. Requiere ROLE_ADMIN."
      },
      "delete": {
        "tags": [
          "Proveedores"
        ],
        "operationId": "eliminar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del proveedor",
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Proveedor eliminado"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Proveedor no encontrado"
          },
          "409": {
            "description": "El proveedor posee órdenes relacionadas"
          }
        },
        "summary": "Eliminar proveedor",
        "description": "Elimina un proveedor existente. Requiere ROLE_ADMIN."
      }
    },
    "/api/promociones/{id}": {
      "get": {
        "tags": [
          "Promociones"
        ],
        "operationId": "obtener_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la promoción",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Promoción encontrada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPromocion"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Promoción no encontrada"
          }
        },
        "summary": "Obtener promoción",
        "description": "Obtiene una promoción por su identificador. Requiere ROLE_ADMIN."
      },
      "put": {
        "tags": [
          "Promociones"
        ],
        "operationId": "actualizar_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la promoción",
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromocionRequest"
              }
            }
          },
          "required": true,
          "description": "Datos actualizados de la promoción"
        },
        "responses": {
          "200": {
            "description": "Promoción actualizada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPromocion"
                }
              }
            }
          },
          "400": {
            "description": "Datos o fechas inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Promoción o producto no encontrado"
          }
        },
        "summary": "Actualizar promoción",
        "description": "Actualiza una promoción existente. Requiere ROLE_ADMIN."
      },
      "delete": {
        "tags": [
          "Promociones"
        ],
        "operationId": "eliminar_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la promoción",
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Promoción eliminada"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Promoción no encontrada"
          }
        },
        "summary": "Eliminar promoción",
        "description": "Elimina una promoción existente. Requiere ROLE_ADMIN."
      }
    },
    "/api/productos/{id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Obtener producto",
        "description": "Busca un producto por su identificador.",
        "operationId": "obtenerProductoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                },
                "example": {
                  "id": 1,
                  "nombre": "Arroz 1 kg",
                  "precio": 1890.0,
                  "stock": 40,
                  "categoria": {
                    "id": 1,
                    "nombre": "Abarrotes"
                  },
                  "_links": {
                    "self": {
                      "href": "http://localhost:8080/api/productos/1"
                    },
                    "productos": {
                      "href": "http://localhost:8080/api/productos"
                    },
                    "categoria": {
                      "href": "http://localhost:8080/api/categorias/1"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Productos"
        ],
        "summary": "Actualizar producto",
        "description": "Reemplaza los datos de un producto existente.",
        "operationId": "actualizarProducto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Producto actualizado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProducto"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "403": {
            "description": "El usuario no posee rol administrador"
          }
        }
      },
      "delete": {
        "tags": [
          "Productos"
        ],
        "summary": "Eliminar producto",
        "description": "Elimina un producto del catálogo.",
        "operationId": "eliminarProducto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Producto eliminado"
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "403": {
            "description": "El usuario no posee rol administrador"
          }
        }
      }
    },
...
    "/api/categorias/{id}": {
      "get": {
        "tags": [
          "Categorías"
        ],
        "summary": "Obtener categoría",
        "operationId": "obtenerCategoriaPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador de la categoría",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Categoría encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Categoria"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          },
          "404": {
            "description": "Categoría no encontrada"
          }
        },
        "description": "Obtiene una categoría por su identificador."
      },
      "put": {
        "tags": [
          "Categorías"
        ],
        "summary": "Actualizar categoría",
        "operationId": "actualizarCategoria",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la categoría",
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Categoria"
              }
            }
          },
          "required": true,
          "description": "Datos actualizados de la categoría"
        },
        "responses": {
          "200": {
            "description": "Categoría actualizada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Categoria"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Categoría no encontrada"
          },
          "409": {
            "description": "Nombre de categoría duplicado"
          }
        },
        "description": "Actualiza una categoría existente. Requiere ROLE_ADMIN."
      },
      "delete": {
        "tags": [
          "Categorías"
        ],
        "summary": "Eliminar categoría",
        "operationId": "eliminarCategoria",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la categoría",
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Categoría eliminada"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Categoría no encontrada"
          },
          "409": {
            "description": "La categoría posee productos relacionados"
          }
        },
        "description": "Elimina una categoría existente. Requiere ROLE_ADMIN."
      }
    },
...
    "/api/productos": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Listar productos",
        "description": "Obtiene todos los productos con enlaces HATEOAS.",
        "operationId": "listarProductos",
        "responses": {
          "200": {
            "description": "Productos obtenidos correctamente",
            "content": {
              "application/hal+json": {
                "example": {
                  "_embedded": {
                    "productoList": [
                      {
                        "id": 1,
                        "nombre": "Arroz 1 kg",
                        "precio": 1890.0,
                        "stock": 40,
                        "_links": {
                          "self": {
                            "href": "http://localhost:8080/api/productos/1"
                          }
                        }
                      }
                    ]
                  },
                  "_links": {
                    "self": {
                      "href": "http://localhost:8080/api/productos"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Productos"
        ],
        "summary": "Crear producto",
        "description": "Agrega un producto al catálogo.",
        "operationId": "guardarProducto",
        "requestBody": {
          "description": "Datos del nuevo producto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              },
              "example": {
                "nombre": "Aceite vegetal 1 L",
                "precio": 2490.0,
                "stock": 30,
                "categoria": {
                  "id": 1
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Producto creado",
            "content": {
              "application/hal+json": {
                "example": {
                  "id": 1,
                  "nombre": "Arroz 1 kg",
                  "precio": 1890.0,
                  "stock": 40,
                  "categoria": {
                    "id": 1,
                    "nombre": "Abarrotes"
                  },
                  "_links": {
                    "self": {
                      "href": "http://localhost:8080/api/productos/1"
                    },
                    "productos": {
                      "href": "http://localhost:8080/api/productos"
                    },
                    "categoria": {
                      "href": "http://localhost:8080/api/categorias/1"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos del producto inválidos"
          },
          "403": {
            "description": "El usuario no posee rol administrador"
          }
        }
      }
    },
...
    "/api/pedidos/{id}": {
      "get": {
        "tags": [
          "Pedidos en línea"
        ],
        "operationId": "obtener_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del pedido",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido encontrado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPedido"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          },
          "404": {
            "description": "Pedido no encontrado"
          }
        },
        "summary": "Obtener pedido",
        "description": "ROLE_CLIENTE solo puede consultar pedidos propios; ROLE_ADMIN puede consultar cualquiera."
      }
    },
...
    "/api/ordenes-compra/{id}": {
      "get": {
        "tags": [
          "Órdenes de compra"
        ],
        "operationId": "obtener_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la orden",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Orden encontrada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrdenCompra"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Orden no encontrada"
          }
        },
        "summary": "Obtener orden de compra",
        "description": "Obtiene una orden por su identificador. Requiere ROLE_ADMIN."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Sucursal": {
        "type": "object",
        "description": "Sucursal de MiniMarket Plus.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "nombre": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "activa": {
            "type": "boolean"
          }
        }
      },
      "EntityModelSucursal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "nombre": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "activa": {
            "type": "boolean"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "Proveedor": {
        "type": "object",
        "description": "Proveedor utilizado para las órdenes de compra.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "nombre": {
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          }
        }
      },
      "EntityModelProveedor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "nombre": {
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "EntityModelPromocion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "nombre": {
            "type": "string"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          },
          "porcentajeDescuento": {
            "type": "number",
            "format": "double"
          },
          "fechaInicio": {
            "type": "string",
            "format": "date"
          },
          "fechaFin": {
            "type": "string",
            "format": "date"
          },
          "activa": {
            "type": "boolean"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Ingrese únicamente el token generado por POST /api/auth/login.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}